+++ private/buf/cmd/buf/command/mod/internal/internal.go
@@ -65,7 +65,7 @@ func BindLSRulesVersion(flagSet *pflag.FlagSet, addr *string, flagName string, a
 	flagSet.StringVar(
 		addr,
 		flagName,
-		"",
+		"", // do not set a default as we need to know if this is unset
 		fmt.Sprintf(
 			"List all the rules for the given configuration version. Implies --%s. Must be one of %s.",
 			allFlagName,