+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Removed the `buf config migrate-v1beta1` command in favor of `buf beta migrate-v1beta1`.
 - Add `buf beta decode` command to decode message with provided image source and message type.
 - Disable `--config` flag for workspaces.
+- Move default config version from `v1beta1` to `v1`.
 
 ## [v1.0.0-rc12] - 2022-02-01